# openapi/openapi.yaml
openapi: 3.0.0
paths:
...
  /auth/register:
    post:
      operationId: AuthController_register
      summary: 用户注册
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterDto'
      responses:
        '201':
          description: 注册成功
        '409':
          description: 用户名或邮箱已存在
      tags: &ref_4
        - auth
  /auth/login:
    post:
      operationId: AuthController_login
      summary: 用户登录
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginDto'
      responses:
        '200':
          description: 登录成功
        '401':
          description: 用户名或密码错误
      tags: *ref_4
  /auth/refresh:
    post:
      operationId: AuthController_refresh
      summary: 刷新令牌
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenDto'
      responses:
        '200':
          description: 刷新成功
        '401':
          description: 刷新令牌无效
      tags: *ref_4
  /auth/logout:
    post:
      operationId: AuthController_logout
      summary: 用户登出
      parameters: []
      responses:
        '204':
          description: ''
      tags: *ref_4
      security:
        - bearer: []
  /auth/me:
    get:
      operationId: AuthController_getCurrentUser
      summary: 获取当前用户信息
      parameters: []
      responses:
        '200':
          description: 获取成功
      tags: *ref_4
      security:
        - bearer: []
    patch:
      operationId: AuthController_updateCurrentUser
      summary: 更新当前用户信息
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMeDto'
      responses:
        '200':
          description: 更新成功
        '400':
          description: 请求参数错误
        '401':
          description: 未认证或令牌无效
      tags: *ref_4
      security:
        - bearer: []
  /auth/users/{userId}:
    get:
      operationId: AuthController_getUserProfileByUserId
      summary: 根据 userId 获取用户信息
      parameters:
        - name: userId
          required: true
          in: path
          description: 用户ID
          schema:
            type: string
      responses:
        '200':
          description: 获取成功
        '401':
          description: 未认证
        '404':
          description: 用户不存在
      tags: *ref_4
      security:
        - bearer: []
...
  /blocks:
    post:
      operationId: BlocksController_create
      summary: 创建块
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBlockDto'
      responses:
        '201':
          description: 创建成功
        '400':
          description: 请求参数错误
        '403':
          description: 没有权限
      tags: &ref_9
        - blocks
      security: &ref_10
        - bearer: []
  /blocks/{blockId}/content:
    patch:
      operationId: BlocksController_updateContent
      summary: 更新块内容
      parameters:
        - name: blockId
          required: true
          in: path
          description: 块ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateBlockDto'
      responses:
        '200':
          description: 更新成功
        '403':
          description: 没有权限
        '404':
          description: 块不存在
      tags: *ref_9
      security: *ref_10
  /blocks/{blockId}/move:
    post:
      operationId: BlocksController_move
      summary: 移动块
      parameters:
        - name: blockId
          required: true
          in: path
          description: 块ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MoveBlockDto'
      responses:
        '200':
          description: 移动成功
        '400':
          description: 移动操作无效
        '403':
          description: 没有权限
        '404':
          description: 块不存在
      tags: *ref_9
      security: *ref_10
  /blocks/{blockId}:
    delete:
      operationId: BlocksController_remove
      summary: 删除块
      parameters:
        - name: blockId
          required: true
          in: path
          description: 块ID
          schema:
            type: string
      responses:
        '200':
          description: 删除成功
        '403':
          description: 没有权限
        '404':
          description: 块不存在
      tags: *ref_9
      security: *ref_10
  /blocks/{blockId}/versions:
    get:
      operationId: BlocksController_getVersions
      summary: 获取块版本历史
      parameters:
        - name: blockId
          required: true
          in: path
          description: 块ID
          schema:
            type: string
        - name: page
          required: false
          in: query
          description: 页码
          schema:
            minimum: 1
            default: 1
            type: number
        - name: pageSize
          required: false
          in: query
          description: 每页数量
          schema:
            minimum: 1
            maximum: 100
            default: 20
            type: number
      responses:
        '200':
          description: 获取成功
        '403':
          description: 没有权限
        '404':
          description: 块不存在
      tags: *ref_9
      security: *ref_10
  /blocks/batch:
    post:
      operationId: BlocksController_batch
      summary: 批量操作块
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchBlockDto'
      responses:
        '200':
          description: 批量操作成功
        '400':
          description: 请求参数错误
        '403':
          description: 没有权限
      tags: *ref_9
      security: *ref_10
...
  /favorites:
    post:
      operationId: FavoritesController_create
      summary: 收藏文档
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFavoriteDto'
      responses:
        '201':
          description: 收藏成功
        '404':
          description: 文档不存在
        '409':
          description: 已经收藏过
      tags: &ref_15
        - favorites
      security: &ref_16
        - bearer: []
    get:
      operationId: FavoritesController_findAll
      summary: 获取收藏列表
      parameters:
        - name: page
          required: false
          in: query
          description: 页码
          schema:
            default: 1
            type: number
        - name: pageSize
          required: false
          in: query
          description: 每页数量
          schema:
            default: 20
            type: number
      responses:
        '200':
          description: 获取成功，含 document 信息
      tags: *ref_15
      security: *ref_16
  /favorites/{docId}:
    delete:
      operationId: FavoritesController_remove
      summary: 取消收藏
      parameters:
        - name: docId
          required: true
          in: path
          description: 文档 ID
          schema:
            type: string
      responses:
        '200':
          description: 已取消收藏
        '404':
          description: 未收藏该文档
      tags: *ref_15
      security: *ref_16
  /comments:
    post:
      operationId: CommentsController_create
      summary: 创建评论（含回复）
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCommentDto'
      responses:
        '201':
          description: 创建成功
        '404':
          description: 文档或父评论不存在
      tags: &ref_17
        - comments
      security: &ref_18
        - bearer: []
    get:
      operationId: CommentsController_findAll
      summary: 获取评论列表
      parameters:
        - name: docId
          required: false
          in: query
          description: 文档 ID
          schema:
            type: string
        - name: blockId
          required: false
          in: query
          description: 块 ID
          schema:
            type: string
        - name: page
          required: false
          in: query
          description: 页码
          schema:
            default: 1
            type: number
        - name: pageSize
          required: false
          in: query
          description: 每页数量
          schema:
            default: 20
            type: number
      responses:
        '200':
          description: 获取成功
      tags: *ref_17
      security: *ref_18
  /comments/{commentId}:
    get:
      operationId: CommentsController_findOne
      summary: 获取评论详情
      parameters:
        - name: commentId
          required: true
          in: path
          description: 评论 ID
          schema:
            type: string
      responses:
        '200':
          description: 获取成功
        '404':
          description: 评论不存在
      tags: *ref_17
      security: *ref_18
    patch:
      operationId: CommentsController_update
      summary: 更新评论
      parameters:
        - name: commentId
          required: true
          in: path
          description: 评论 ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCommentDto'
      responses:
        '200':
          description: 更新成功
        '403':
          description: 只能修改自己的评论
      tags: *ref_17
      security: *ref_18
    delete:
      operationId: CommentsController_remove
      summary: 删除评论
      parameters:
        - name: commentId
          required: true
          in: path
          description: 评论 ID
          schema:
            type: string
      responses:
        '200':
          description: 删除成功（软删除）
        '403':
          description: 只能删除自己的评论
      tags: *ref_17
      security: *ref_18
...
info:
  title: 知识库 API
  description: 个人知识库系统 API 文档
  version: '1.0'
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    RegisterDto:
      type: object
      properties:
        username:
          type: string
          description: 用户名
          example: john_doe
        email:
          type: string
          description: 邮箱
          example: john@example.com
        password:
          type: string
          description: 密码
          example: SecurePass123!
        displayName:
          type: string
          description: 显示名称
          example: John Doe
      required:
        - username
        - email
        - password
    LoginDto:
      type: object
      properties:
        emailOrUsername:
          type: string
          description: 邮箱或用户名
          example: john@example.com
        password:
          type: string
          description: 密码
          example: SecurePass123!
      required:
        - emailOrUsername
        - password
    RefreshTokenDto:
      type: object
      properties:
        refreshToken:
          type: string
          description: 刷新令牌
      required:
        - refreshToken
    UpdateMeDto:
      type: object
      properties:
        displayName:
          type: object
          description: 显示名称
          example: John Doe
        avatar:
          type: object
          description: 头像地址
          example: https://example.com/avatar.jpg
        bio:
          type: object
          description: 个人简介
          example: 这是我的简介
...
    CreateBlockDto:
      type: object
      properties:
        docId:
          type: string
          description: 文档ID
          example: doc_1234567890_abc123
        type:
          type: string
          description: 块类型
          example: paragraph
        payload:
          type: object
          description: 块内容（JSON格式）
          example:
            text: 块内容
        parentId:
          type: string
          description: 父块ID（根块为空）
          example: b_1234567890_abc123
        sortKey:
          type: string
          description: 排序键
          example: '0'
        indent:
          type: number
          description: 缩进级别
          example: 0
          default: 0
        collapsed:
          type: boolean
          description: 是否折叠
          example: false
          default: false
        createVersion:
          type: boolean
          description: 是否立即创建文档版本
          example: true
          default: true
      required:
        - docId
        - type
        - payload
    UpdateBlockDto:
      type: object
      properties:
        payload:
          type: object
          description: 块内容（JSON格式）
          example:
            text: 更新的块内容
        plainText:
          type: string
          description: 纯文本内容（用于搜索）
          example: 更新的块内容
        createVersion:
          type: boolean
          description: 是否立即创建文档版本
          example: true
          default: true
      required:
        - payload
    MoveBlockDto:
      type: object
      properties:
        parentId:
          type: string
          description: 目标父块ID
          example: b_1234567890_abc123
        sortKey:
          type: string
          description: 排序键
          example: '0.5'
        indent:
          type: number
          description: 缩进级别
          example: 0
          default: 0
        createVersion:
          type: boolean
          description: 是否立即创建文档版本
          example: true
          default: true
      required:
        - parentId
        - sortKey
    BatchBlockDto:
      type: object
      properties:
        docId:
          type: string
          description: 文档ID
          example: doc_1234567890_abc123
        operations:
          type: array
          description: 批量操作列表
          items:
            oneOf:
              - $ref: '#/components/schemas/BatchCreateOperation'
              - $ref: '#/components/schemas/BatchUpdateOperation'
              - $ref: '#/components/schemas/BatchDeleteOperation'
              - $ref: '#/components/schemas/BatchMoveOperation'
        createVersion:
          type: boolean
          description: 是否立即创建文档版本
          example: true
          default: true
      required:
        - docId
        - operations
...
    CreateFavoriteDto:
      type: object
      properties:
        docId:
          type: string
          description: 文档 ID
      required:
        - docId
    CreateCommentDto:
      type: object
      properties:
        docId:
          type: string
          description: 文档 ID
        blockId:
          type: string
          description: 块 ID（评论挂在块上时）
        content:
          type: string
          description: 评论内容
        mentions:
          description: '@ 提及的用户 ID 列表'
          type: array
          items:
            type: string
        parentCommentId:
          type: string
          description: 父评论 ID（回复时）
      required:
        - docId
        - content
    UpdateCommentDto:
      type: object
      properties:
        content:
          type: string
          description: 评论内容
      required:
        - content
...